@@ -39,16 +39,16 @@
     "bootstrap": "^5.3.2",
     "case-sensitive-paths-webpack-plugin": "^2.4.0",
     "classnames": "^2.5.1",
-    "es6-promise": "^3.3.1",
+    "es6-promise": "^4.2.8",
     "eslint-plugin-babel": "^3.3.0",
     "eslint-plugin-react": "^7.33.2",
     "file-loader": "^6.2.0",
     "html-webpack-plugin": "^5.3.2",
     "imports-loader": "^3.0.0",
-    "jquery": "^3.6.0",
+    "jquery": "^3.7.1",
     "lodash": "^4.14.0",
     "postcss-loader": "^6.1.1",
-    "prop-types": "^15.7.2",
+    "prop-types": "^15.8.1",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
     "react-redux": "^7.2.5",
@@ -64,10 +64,10 @@
     "style-loader": "^3.3.0",
     "url-loader": "^4.1.1",
     "watchify": "^4.0.0",
-    "webpack": "^5.51.1",
-    "webpack-bundle-tracker": "1.3.0",
-    "webpack-cli": "^4.8.0",
-    "webpack-dev-server": "^4.2.1",
+    "webpack": "^5.90.1",
+    "webpack-bundle-tracker": "^3.0.1",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1",
     "axios": "^1.6.7",
     "whatwg-fetch": "^3.6.20"
   },
