@@ -1,47 +1,42 @@
 import json
-from test_plus.test import TestCase
-from rest_framework.test import APIClient
-
-from django.test import RequestFactory
+from django.test import TestCase, Client
 from django.urls import reverse
-
 from saas_app.users.models import User
 
 
 class BaseUserTestCase(TestCase):
     def setUp(self):
-        self.user = self.make_user()
-        self.factory = RequestFactory()
-        self.client = APIClient()
+        # Assuming `make_user` is a helper method to create a user instance
+        # If not, replace this with the appropriate User creation logic
+        self.user = User.objects.create_user(username='testuser', email='test@example.com', password='password')
+        self.client = Client()
 
     def login(self):
-        super(BaseUserTestCase, self).login(
-            username=self.user.username, password="password"
-        )
+        self.client.force_login(self.user)
 
 
 class CurrentUserViewTestCase(BaseUserTestCase):
 
-    url = "api:v1:user"
+    url = reverse("api:v1:user")  # Moved reverse call to class attribute for optimization
 
     def test_needs_login(self):
-        resp = self.client.get(reverse(self.url))
-        self.assertEqual(resp.status_code, 403)
+        response = self.client.get(self.url)
+        self.assertEqual(response.status_code, 403)
 
     def test_get(self):
         self.login()
-        resp = self.client.get(reverse(self.url))
-        data = json.loads(resp.content.decode("utf-8"))
+        response = self.client.get(self.url)
+        data = json.loads(response.content.decode("utf-8"))
         self.assertEqual(data["email"], self.user.email)
 
     def test_put_valid(self):
         self.login()
-        resp = self.client.put(reverse(self.url), {"name": "foo"}, format="json")
-        self.assertEqual(resp.status_code, 200)
+        response = self.client.put(self.url, {"name": "foo"}, content_type='application/json')
+        self.assertEqual(response.status_code, 200)
         user = User.objects.get(pk=self.user.pk)
-        self.assertEquals(user.name, "foo")
+        self.assertEqual(user.name, "foo")
 
     def test_put_invalid(self):
         self.login()
-        resp = self.client.put(reverse(self.url), {})
-        self.assertEqual(resp.status_code, 400)
+        response = self.client.put(self.url, {}, content_type='application/json')
+        self.assertEqual(response.status_code, 400)
